LayoutTests/inspector/dom/content-node-region-info-expected.txt

Testing that the DOMTreeManager.getNodeContentFlowInfo returns the containing regions.


Selector: #flow
Region flow name: none
Content flow name: none
Regions count: N/A
Regions: N/A

Selector: .contentNode1
Region flow name: none
Content flow name: none
Regions count: N/A
Regions: N/A

Selector: .contentNode2
Region flow name: none
Content flow name: none
Regions count: N/A
Regions: N/A

Selector: body
Region flow name: none
Content flow name: none
Regions count: N/A
Regions: N/A

Selector: #region1
Region flow name: none
Content flow name: none
Regions count: N/A
Regions: N/A

Selector: #region2
Region flow name: none
Content flow name: none
Regions count: N/A
Regions: N/A
